+++ src/WorkspaceManager.ts
@@ -149,7 +149,7 @@ export class WorkspaceManager implements vscode.Disposable {
       configuration.getEnableStrictPattern(),
       configuration.getGoogleTestTreatGMockWarningAs(),
       configuration.getGoogleTestGMockVerbose(),
-      true,
+      false,
     );
 
     this._disposables.push(